Add in appointment reschedule confirmation and pull out appointment confirmation generation
